Add responsive and change logo
The login page is now responsive with the home page.

The login page now has a logo in the bottom right-hand corner. Only the logo appears when you're on your phone.

The navigation bar can now be reduced (but not enlarged again for the moment, unless the page reloads). The navigation bar is now also responsive, disappearing on the phone and appearing when a button is clicked.